refactore login and signup into separate views
